Small changes, renamed test files [#158039176]

+++ src/Housing.Selection.Library/RoomSearchViewModel.cs
@@ -4,7 +4,7 @@
     {
         public string Location { get; set; }
         public string Batch { get; set; }
-        public int BatchMinimumPercentage { get; set; }
+        public double BatchMinimumPercentage { get; set; }
         public char Gender { get; set; }
         public bool IsCompletelyUnassigned { get; set; }
     }
